Magistracy/теория_чисел/лабораторные_работы/лабораторная_работа_1/pyproject.toml
[tool.poetry]
name = "laboratory-work-1"
version = "1.0.0"
description = ""
authors = ["Deniz <68436196+Denzi33@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
pytest = "^8.3.3"
Sphinx = { version = "^4.2.0", optional = true }
sphinx-rtd-theme = { version = "^1.0.0", optional = true }
sphinxcontrib-napoleon = { version = "^0.7", optional = true }

[tool.poetry.extras]
docs = ["Sphinx", "sphinx-rtd-theme", "sphinxcontrib-napoleon"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
